refactor: move navigation bar/footer assertion to utils
