+++ web/frontend/tests/integration/footer.spec.ts
@@ -3,7 +3,3 @@ import { default as i18n } from 'i18next';
 import { initI18n } from './utils';
 
 initI18n();
-
-export async function getFooter(page) {
-  return await page.locator('//*[@id="root"]/div[1]/div[4]/div[1]/footer');
-}
