+++ web/frontend/tests/integration/homepage.spec.ts
@@ -1,17 +1,11 @@
 import { test, expect } from '@playwright/test';
-import { getFooter } from './footer.spec';
+import { assertNavigationPresent, assertFooterPresent } from './utils';
 
 test('Assert homepage has correct title', async({ page }) => {
   await page.goto(process.env.FRONT_END_URL);
   await expect(page).toHaveTitle(/D-Voting/);
 });
 
-test('Assert navigation bar is present', async({ page }) => {
-  await page.goto(process.env.FRONT_END_URL);
-  await expect(page.getByRole('navigation')).toBeVisible();
-});
+test('Assert navigation bar is present', async({ page }) => assertNavigationPresent(page));
 
-test('Assert footer is present', async({ page }) => {
-  await page.goto(process.env.FRONT_END_URL);
-  await expect(await getFooter(page)).toBeVisible();
-});
+test('Assert footer is present', async({ page }) => assertFooterPresent(page));
